Create tables prior to descriptor update
AIPS UV files need to be created in a particular sequence for
the visibility buffer to be available for writing. This appears to arises
from the fact that tables such as antenna, frequency and source
must be created before the descriptor is updated, to avoid it and other
data structures from getting into an inconsistent state.