Fix export in continuum_pipeline.py (#190)
Within the continuum pipeline, katdal selection is used for two purposes:

1. To condition AIPS UV files with observation dimensions
   and parameters.
2. Select a subset of the observation.

The two do not necessarily coincide. Within the pipeline, we often
want to write all targets to the AIPS SU table into an AIPS UV file,
but only want to write one scans worth of data. A katdal selection
for the scan results in only one target.

This commit properly implements the (1) conditioning and
(2) data selection in the correct places.