Improve UVFacade and ImageFacade robustness
Once we're done with a UV or Image object, it's necessary to manually
delete it in order to free up internal resources. See SR-962 for the
history of this.

By deleting these objects, our Facade is either missing an attribute
or it's set to None. Resolve this by provide a `uv` and `img` property
on UVFacade's and ImageFacade's respectively, which

* Returns the internal UV or Image object, if they exist.
* Re-creates it if it does not.