Correct scan start and end times
https://github.com/ska-sa/katsdppipelines/pull/202 chunked up each
scan's data by time, but the logic for updating the AIPS scan
time and visibility boundaries was not updated commensurately.

In effect, the starting time and starting visibility of the last
time chunk was used to set the scan boundaries, rather than
those of the entire scan. This commit corrects this.