Allow user defined configuration at runtime
The continuum_pipeline.py script is self contained so there is no need to define AIPS disks in advance. Instead let a disk be set up and defined at runtime via the --workdir argument. Also output the Obit task logs to the workdir so that that can be inspected on mounted volumes.