Convert AIPS SN gains into numpy arrays and export them to telstate
The function _massage_gains will take an AIPS SN AIPSTable object
and convert it to an ndarray of shape (nif,npol,nants).
Undefined gaines are encoded as a float32 view of the bytestring
'INDEF'. These are checked for and assigned to np.nan if their weights
are also less than or equal to 0.