Add tests for export of SN tables to telstate.
This generates an SN table and populates it with a few corner case
variations of nans in various locations in the real and imaginary
parts of the visibilities. This test also brough out a bug in
AIPSTable that meant that newly created tables were not properly
having defaults set so I fixed this here as well.