Add tests of self-calibration in the pipeline
This adds tests that the gains in telstate are as expected
after a run of the pipeline on a simulated dataset containing
a 10Jy source at the phase-center. It checks that the telstate
keys and timestamps are correctly set and exist as they should
for phase+amp, phase-only and no self-cal. This test turned up a bug
in MockDataset that incorrectly set the timestamps to the start of
a dump rather than the centre so I fixed that here as well.