Add an array reference antenna to the mock dataset
Since ska-sa/katdal#252 the tests for katsdpcontim have been failing.
This is because the new method of calculating (u,v,w) coordinates
relies on an array reference antenna sensor which was not provided
by MockDataset. So we add a reference antenna in the same way
it is done in h5datav3 (i.e. use the first antenna as the array ref.).