Add a new class `Pipeline` that is a base class for `ContinuumPipeline`
The idea here is that `Pipeline` is a base class for different kinds of
imaging pipeline (eg. offline vs. online). It provides a method `_execute`
that must be defined by any derived class that uses it and will do the
actual work of the desired pipeline.

The new derived class `ContinuumPipeline` is (as it used to be) the
pipeline for the online MeerKAT SDP Continuum Imager and requires a
`TelescopeState` for export of solutions (the offline one will not)
and is set up for automation.

I've also added a new kwarg called `prtlv` to specify the amount of
logging from Obit tasks (in preparation for the Offline pipeline
where the user can select it directly). I've also set things up so that
when an exception is raised by the user supplied `_execute` method,
the desired cleanup of the aipsdisk will always happen before exiting.