Add --reuse option to image data from aipsdisk without archive retrieval.
The --reuse parameter takes the disk location of an aipsdisk. If this
contains a <CB_ID>.merge file, this is assumed to have been produced
during a previous run of the pipeline and the data are not copied to
local disk from the archive. This allows many runs of the imager with
different parameters on the same dataset without the time overhead
of downloading and converting the data.

I've separated out the code that does selection on the katdal object
and infers the aipsdisk file names into a separate method in `Pipeline`.
I then search for the inferred 'merge' filename in the aipsdisk for
reusing and then pass this into mfimage.