Tidy things up a bit.
Now that `continuum_pipeline.py` exists only for online use, change
some of the options to reflect this. I've separated out configuration
.yaml files specifically for the online case and hard-coded their locations
in the docker image and removed the command line option to specify them.
The script no clobbers all of the files in the aipsdisk by default and
I've fixed a small bug in the `clobber` method.

Also some small tidying up in the `continuum_image.py` script as well.