Add default parameters inferred from katdal object.
Infer the refant from the cal_refant and the channel averaging
so we should always average to 1K channels for continuum.

Delete relevant parameters from the configuration defaults.

Also move the maxRuntime code into the infer_defaults function.

Also Fix broken baseline averaging parameter. (This should be
the same as what we use as the self-calibration interval.)