Truncate AIPS path names to 12 characters
Also- ensure that the UV header 'SOURCE' keyword has 16 characters.

The AIPS file names of imaged targets are derived from their SOURCE
keyword in the input AIPS UV header. Source names were being truncated
to 16 characters, which is the required length of the SOURCE keyword
in an AIPS UV header, via the `normalise_target_name` function. AIPS
file names have a maximum length of 12 characters. While some Obit
tasks can handle this descrepancy (i.e. MFImage will truncate filenames
derived from the SOURCE parameter to 12 characters), others cannot
(i.e. TabCopy) and require that filenames (derived from outName and
inName parameters) are a maximum of 12 characters.

The workaround is to require that all SOURCE header keywords (derived
from katpoint target names) are truncated to 12 characters, then
pad this with spaces up to the 16 characters required. This also ensures
that **unique** filenames will be created when there are multiple
targets with the same first 12 characters in their name. The
output FITS filenames are derived from the katpoint descriptions
and remain unchanged.

Also update a test to include a longer filename for data that is passed
to TabCopy.