Update Obit to revision 616.
Revision 616 has all the updates needed for python3 compatibility but
still has backwards compatibility with python2. The Makefiles have
changed a little in this revision so update the patch to ensure that
it compiles inside the docker image. Also the IODesc parameter on
tables is no longer being used, its functionality is now just the same
as the Desc parameter- so change all use of IODesc to Desc in aips_table.py.