Fix bug in uv.Desc.Dict introduced in Obit r616
Online testing of this branch failed because one of the scans
in the test had completely flagged data. The baseline averager
throws away all of the data in a scan if it is completely flagged.
Obit r616 incorrectly reports the number of visibility rows in the
uv data output from a completely flagged scan (after baseline averaging),
causing the pipeline to try and read data that isn't there.

The fix is to rather count the number of visibilities from the attached
NX table that is produced by the baseline averaging. (See AIPS MEMO 117 for
a description of NX tables). This should be robust as the NX table is
produced by looping over the visibilities in the input data and counting
them directly.

Also add a test, which ensures that completely flagged data successfully
runs through the process of conversion to AIPS UV with baseline averaging.