Update PNG generation to use similar method as in katsdpimager.
This is **mostly** a cut and paste of the code in render.py from katsdpimager.
The differences are predominatly the result of the FREQ axis in the continuum images
being differently named ('SPECLNMF') in Obit images, and it being the 3rd image axis
rather than the 4th. The specific changes are:
* I have added a trim parameter that defines the fraction of pixels at the
edges of the images to omit. This does the same job as the trim parameter
as it was previously defined in this module but uses matplotlib to do the work. One
minor side effect of this is that zscale limits are now computed from the full image
rather than the trimmed image. This seems to make little difference in practice.
* Move the reading of the FITS 'BUNIT' card out of _plot. Now pass the 2d data array
and the 'BUNIT' value in from outside. This means that the :class:`HDUList` object
doesn't need to be passed in and therefore avoids required assumptions about the axis
ordering of the FITS image.
* Remove the computation of the frequency of the image from _plot, this can be brought
in using the 'caption' parameter.
* Add a new 'slices' parameter to write_image which selects the desired image plane to
plot. This parameter is passed directly to :Class:`WCSAxes` in _prepare_axes, it is
further used to get the relevant data from the :class:`HDUList` object to pass into _plot
(so that only a 2d data array is passed to _plot).