Fix up failing tests.
The pipeline tests for a corner case involving multiple targets
of the same name were producing images with garbled parameters
in the frequency axis - due to the number of IFs in the dataset
being larger than the number of channels. It seems Obit just ploughs
on in this case rather than dying with error or even warning
about a potential problem. The WCS object constructor didn't
like these broken FITS header.

Fix it by generating datasets with sufficient channels.