Add azimuth and elevation sensors to MockDataSet.
This allows the az,el and parangle methods on a MockDataSet to function.
These are needed for MeerKAT+ simulations, which use parralactic angles
to interpolate the primary beam and to calculate linear feed rotation.
There are also some other small modifications, one to correct an error
in the activity sensor of the refant and another to add a list of antenna
names to the initial dataset selection (as this is what a real katdal
DataSet does.