Increase the default time chunk for reading data to 20 dumps.
Also expose setting the time chunks at the script level.
The main reason for doing this is that my MeerKAT+ simulation parallelises
nicely over time and big gains can be had by simulating many dumps at
once. On top of this the previous default (4) was very low even for real
data, and things are slightly quicker to read from the archive (maximum 5%)
with bigger chunks.

Memory usage isn't a problem. For the MeerKAT pipeline case, using the
continuum stream with 1024 channels reading the data used a maximum
of 1GB of memory, so now it will use 5GB. In the MeerKAT+ world the data
size in the continuum stream should increase by a factor of 3.4 per dump
(2 times the channels and 1.7 times the baselines), my simulations
show that memory usage in this case is ~17GB as expected.

I've also increased the default for warning about the memory usage
to 8GB (calculated from the size of the visibilities alone without
any memory overheads).