Export available self-calibration solutions
In the event that the imager times out before the self-calibration
cycles are complete, make sure that calibration solutions that are
available are still exported.

This is done by assuming the AIPS SN tables attached to the UV data
per target are added with sequential version numbers starting from 1
for each self-calibration cycle. The solution is then to accept the
phase-only self-cal solution table as the one which is the minimum
of the MFImage parameter `maxPSCLoop` (the commanded number of phase
only self cal cycles) and the actual maximum version of the attached
SN tables. For amp+phase solutions take the minimum of the value
of `maxPSCLoop + maxASCLoop` and the maximum SN table version
(`maxASCLoop` is the commanded number of amp+phase self-cal cycles
- and these are only done AFTER the phase-only self-cal cycles is
complete). Only export the amp+phase self cal table if `maxASCLoop > 0`
(i.e. amp+phase self cal was requested) and if the image timed out
AFTER completing all of its phase-only self cal cycles.