Reset list of AIPS disks when exiting Obit context.
The tests I added in the previous commit broke the nosetests. This
turned out to be due to the Obit context used by the pipeline
preserving some state after shutdown. In particular the list of AIPS
disks ended up having over 100 entries due to the frequent opening
and closing of different Obit contexts during the tests. This problem
should only affect the tests since the normal use case is to only
open a single Obit context for each run of the pipeline.

Fix the problem by resetting the list of AIPS disks manually each time
an Obit context is closed.

In reality this is an Obit bug but it might take some effort to fix
that, and the current solution works just fine.