Place each QA image product in its own directory
Instead of grouping all the QA images of a particular type (e.g MEAN,
RMS) into a single directory containing all observed targets for the
observation each image (per target) is now written to its own directory with an
individual metadata file.

The directories are labelled according to the scheme
CBID_continuum_image_START_TIME_TARGET_NAME_RUN_TYPE, where
START_TIME is the time the pipeline run begins, the RUN is the
first target index in the capture block and the type is one of 'QA',
'RMS', 'BKG' and PB corresponding to the QA report, the rms image, the
mean image and the primary beam corrected image respectively.