Ensure the returned array from GetPlane is copied.
Obit R632 contains a plug for a memory leak that has caused
an issue in img_facade. The GetPlane method needs to return a
copy of the image array otherwise if the original image object
is freed from memory, future calls to the array data can try
and look for it at a freed memory address and cause a segfault.
(Why this ever worked before is a mystery!)