Move radius of outlier facets to 1.75 degrees.
I'm doing this to be safe, some further testing with the 2 degree radius
used in the previous commit showed that in certain cases the runtime
can increase to nearly twice the observation length.
I've run simulations putting a 30 Jy point source at 1.75 degrees from
the phase center and there are no residuals in the image area if such a
source is not deconvolved (The brightest peak in SUMSS/NVSS is
~30 Jy/beam). It should therefore be safe to use a 1.75 degree redius for
outlier facets (thereby potentially missing <30Jy sources beyond 1.75 degrees).