Add machinery for selecting band in the offline and online scripts.
UHF-, L-, and S-band configurations are stored in the conf directory.
The files with _MKAT appended are the online imager defaults and those
without are the offline defaults.
The band is selected from the input katdal DataSet using the band
parameter of the selected SpectralWindow object.
This will default to Obit defaults if the assumed input .yaml file
isn't found.
In the offline script 'continuum_image.py', the input parameters
are user configurable using the --mfimage-config and --uvblavg-config
options.