Update L-band parameters for offline imaging.
The offline parameters now represent a 'better case' for full continuum
imaging rather than the online parameters that are optimised for running
within twice the observing time and for continuum subtraction in spectral.

The updated parameters include a larger image area, more frequency
sub-bands, a larger area to search for outliers in NVSS/SUMSS and a
shorter self-calibration interval. The offline imager can anyhow be
used to mimic the online one by specifying the mfimage_ and
uvblavg_L_MKAT.yaml files using the --mfimage-config and
--uvblavg-config parameters in continuum_image.py.