Update UHF parameters for online and offline imaging.
The new online parameters are those described and tested in SPR1-879.
The Field of View is increased by a factor 1.6 to account for the
change in center frequency between the L-band and UHF-band (1284 vs
815 MHz). UHF data generally take longer to image than L-band for an
equivalent data size - this is because on average the radio source
population in brighter (alpha=-0.7) and the sky area is greater. To
deal with this I have increased the fractional bandwidth of the coarse
frequency planes in imaging (maxFBW) to 0.11 from its L-band value of
0.075. This results in 7 coarse frequency planes in UHF band (rather
than the 10 in L-band) and gives a nice speedup in in imager. I have
tested that the CLEAN components with just 7 subbands results in adequate
continuum subtraction (it does). I have also increase the number of
CC iterations to 60000 to account for the brighter source population
and larger sky area. The flux limits for imaging remain the same since
the UHF SEFD of MeerKAT is roughly equivalent to L-band.

The new offline parameters are similarly scaled versions of the offline
L-band parameters - although I keep the maxFBW parameter at its
'sweet spot' of 0.05 (resulting in 15 coarse frequency planes for UHF).