Update S-band parameters for online and offline imaging
I have written up a discussion of the choices for the online parameters
in a comment in SPR1-879 dated 30/09/2021.

The offline parameter choices were arrived at through experimentation.
These mainly consists of a deeper CLEAN (40 microJy/beam instead of 50)
and a narrower maxFBW (0.03 instead of 0.05) to create 10 coarse frequency
planes in S4.
I have increased the self calibration interval to 1 minute for offline
since with the L-band default of 30 seconds the S/N on the solutions
was too low to get reliable gains.