Improve contrast of PNG images.
The contrast of the raw and primary beam corrected PNG images output by the pipeline
was using the default value of 0.02 from the zscale algorithm in katsdpimageutils.
This updates the contrast to 0.5 for the raw MFImage output and to 0.3 for the
PB corrected image output. The mean and rms images are fine with the default value and
I've left them unchanged.
The improved contrast makes it easier to see features in the images that can be useful
for QA appraisal.
Some example of the effect of these changes are shown in JIRA ticket SPR1-1296