Exit gracefully if 100% of the data is flagged
This is to address SPR1-1830, currently if all the visibilities are
flagged the online pipeline throws an exception. This PR checks that
there a non-zero number of unflagged visibilities in merged dataset
before proceeding to imaging. If no unflagged visibilities exist the
pipeline will log and error and terminate without attempt to image an
empty dataset.