Move common command line arguments to Util.
Also a general tidy up of the command line arguments.
Some of the common arguments were not exposed in the
continuum_pipeline.py scripts. These are now dealt with
as necessary.

I've slightly changed the way the parameter files work.
The user can either specify a directory containg the files
the files will be looked for with default names. Otherwise
specific yaml files can be passed.

Also a very minor flake8 cleanup of setup.py.