Add a script `uv_export.py` that generates AIPS UV from MVF
Modify the logging handler a little so that the script
can print the header of the resulting UV file to the log.
Also, change the `export_and_merge_scans` function to only
set up a `uv_merge_path` if the child class hasn't already
defined it. This allows a user definable `uv_merge_path`.