Change nvispio parameter to from 10240 to 1024 everywhere
The nvispio parameter defines the size of a memory buffer that fills before
flushing data to disk when creating AIPS UV. The size 10240 can cause buffer
overflows when there are two many channels to process at in a single baseline
(ie. in 32K mode). Setting this to 1024 prevents this problem, and has no
apparent penalty in terms of write speed for AIPS UV data.