Add catalogue for outlier facets and move the parameter files
The FITS catalogue used to come from the Obit distribution, which had a well defined
location inside the docker container, now it is better to include it in this
package so it can be located easily in the scripts.

Separate the FITS catalogue and the yaml config files by putting them in separate
directories under conf.