Clean up the configuration setup in the scripts.
Much of the config was shared between the different scripts so I've
moved it into a function in utils that is called by all of them. Also
update the post_process_args function to add defaults for args that
might not be used by the scripts for are needed for configuration.