Clean up the config files
Clean up the text in the config files to accurately reflect the changes I've made to them.

Also change maxFBW to -1.0 for narrowband, this will cause the default number of IFs in
narrowband to be the same as the coarse frequency planes. This is acceptable since using
the fractional bandwidth to define the size of the planes isn;t as critical for shorter
bandwidths.