Get CUDA samples from the NVIDA github repo.
As of CUDA 11.6 the samples are no longer available in the toolkit
so retrieve them from NVIDIA's github repo instead. The headers are
placed in /usr/local/cuda/samples/inc which is where the GPU Obit
build modified by 'obit.patch' looks for them.
This should make things easier to maintain for newer versions of CUDA
in future.