Initialise QUAL as an integer in AIPS SU Table.
This is a long-standin bug in katsdpcontim that breaks things with
Python 3.10. The swig code to define AIPS tables requires an integer
for the QUAL field, but it has always been initialised as a float.

This supresses:
DeprecationWarning: an integer is required (got type float).  Implicit
conversion to integers using __int__ is deprecated, and may be removed
in a future version of Python.