SPP1-171 Support MKAT extension dish names
MKAT extension dishes will be labeled with names beginning with 'e' and
3 digit identifying numbers e.g. 'e121' while SKA dishes with be labeled
with names beginning with 's' and 4 digit identifying numbers e.g.
's0001'.
This PR enables the pipeline to accept dish names of all three types
used in the MeerKAT+ and SKA array and to correctly convert between
katdal dish names and aips antenna numbers.