SPP1-171 Support arbitrary dish names
This PR removes the need to conform to a predefined name scheme in the
katdal dataset when converting between katdal mvfv4 datasets and AIPS
files. Previously the AIPS antenna number depended on the antenna number
and naming scheme, whereas going forward the AIPS antenna number will
depend only on its position in the list of antennas included in the
observation. This change will simplify support for any naming changes
that occur in the future.